[refactor] Pay Class 분리
